eliminar 2.0

+++ ASGARD/ClientApp/src/app/components-catalogos/form-clasificacion/form-clasificacion.component.ts
@@ -1,6 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Input } from '@angular/core';
 import { CatalogosService } from './../../services/catalogos.service';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { ActivatedRoute, Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'form-clasificacion',
@@ -8,10 +10,11 @@ import { FormGroup, FormControl, Validators } from '@angular/forms';
   styleUrls: ['./form-clasificacion.component.css']
 })
 export class FormClasificacionComponent implements OnInit {
-  clasificaciones: any;
-  p: number = 1;
+  @Input() clasificaciones: any;
+  //p: number = 1;
   clasificacion: FormGroup;
-  constructor(private catalogosServices: CatalogosService) {
+  display = 'none';
+  constructor(private catalogosServices: CatalogosService,  private router: Router, private activateRoute: ActivatedRoute) {
     this.clasificacion = new FormGroup({
       'idclasificacion': new FormControl("0"),
       'clasificacion': new FormControl("", [Validators.required]),
@@ -25,28 +28,91 @@ export class FormClasificacionComponent implements OnInit {
   }
 
   ngOnInit() {
-    this.catalogosServices.getClasificacion().subscribe(data => { this.clasificaciones = data })
+    this.catalogosServices.getClasificacion().subscribe(data =>  this.clasificaciones = data )
 
+  }
+
+  open() {
+    //limpia cache
+    this.clasificacion.controls["idclasificacion"].setValue("0");
+    this.clasificacion.controls["clasificacion"].setValue("");
+    this.clasificacion.controls["correlativo"].setValue("");
+    this.clasificacion.controls["descripcion"].setValue("");
+    
+    this.display = 'block';
+  }
+  close() {
+    this.display = 'none';
+  }
+
+  eliminar(idclasificacion) {
+    Swal.fire({
+      title: '¿Estas seguro de eliminar este registro?',
+      text: "No podras revertir esta accion!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, eliminar!'
+    }).then((result) => {
+      if (result.value) {
+        this.catalogosServices.eliminarCasificacion(idclasificacion).subscribe(data => {
+          Swal.fire(
+            'Registro eliminado!',
+            'Tu archivo ha sido eliminado con exito.',
+            'success'
+          )
+          this.catalogosServices.getClasificacion().subscribe(
+            data => { this.clasificacion = data }
+          );
+        });
+
+      }
+    })
   }
 
   guardarDatos() {
 
     if (this.clasificacion.valid == true) {
-      this.catalogosServices.guardarClasificacion(this.clasificacion.value).subscribe(data => { })
+      this.catalogosServices.guardarClasificacion(this.clasificacion.value).subscribe(data => { });
+      Swal.fire({
+        position: 'top-end',
+        icon: 'success',
+        title: 'Dato Guardado con exito',
+        showConfirmButton: false,
+        timer: 3000
+      })
+      this.clasificacion.controls["idclasificacion"].setValue("0");
+      this.clasificacion.controls["clasificacion"].setValue("");
+      this.clasificacion.controls["correlativo"].setValue("");
+      this.clasificacion.controls["descripcion"].setValue("");
+      //this.router.navigate(["/form-marca"])
+
+      this.display = 'none';
+      this.catalogosServices.getClasificacion().subscribe(res => this.clasificacion = res);
+
     }
   }
 
+  modif(id) {
 
-  eliminar(idclasificacion) {
-    //if (confirm("desea eliminar el registro?") == true) {
-      this.catalogosServices.eliminarCasificacion(idclasificacion).subscribe(data => {
-        this.catalogosServices.getClasificacion().subscribe(
-          data => { this.clasificacion = data }
-        );
+    this.display = 'block';
+    this.catalogosServices.RecuperarClasificacion(id).subscribe(data => {
 
-      });
-    //}
+      this.clasificacion.controls["idclasificacion"].setValue(data.idclasificacion);
+      this.clasificacion.controls["clasificacion"].setValue(data.clasificacion);
+      this.clasificacion.controls["correlativo"].setValue(data.correlativo);
+      this.clasificacion.controls["descripcion"].setValue(data.descripcion);
+     
+
+      //this.marca.controls["idMarca"].setValue(data.idMarca);
+      //this.marca.controls["marc"].setValue("123");
+      //this.marca.controls["descripcion"].setValue(data.descripcion);
+
+    });
   }
 
-
+  buscar(buscador) {
+    this.catalogosServices.buscarClasificacion(buscador.value).subscribe(res => this.clasificacion = res);
+  }
 }
